+++ core/src/main/java/io/anemos/metastore/core/proto/profile/ProfileAllowAdd.java
@@ -2,6 +2,6 @@ package io.anemos.metastore.core.proto.profile;
 
 public class ProfileAllowAdd extends ProfileAllowAddBase {
   public ProfileAllowAdd() {
-    super("allow:add");
+    super("allow:add", false);
   }
 }