+++ server/src/main/java/io/anemos/metastore/RegistryService.java
@@ -213,7 +213,7 @@ public class RegistryService extends RegistryGrpc.RegistryImplBase {
         profile = new ProfileAllowAdd();
     }
 
-    return profile.validate(results.getReport());
+    return profile.validate(results.createProto());
   }
 
   @Override